Impl Competitor Automatic Operating System
- 경쟁사 자동 운영 시스템 구현 완료
- 기존에는 랜덤으로 손님에게 리뷰를 받았으나, 탐색+학습 방식으로 경쟁사가 손님에게 최고 평점을 받았던 조합을 내거나 새로운 조합을 시도하도록 변경하였다.
- 손님에게 받는 평점이 소수점 첫째자리에서 반올림되지 않는 문제 해결 완료
- 손님의 취향과 일치하지 않는 샌드위치 재료 개수를 세는 함수(CountUnpreferredIngr) 버그 수정

+++ Source/CookYourWay/GameInstance/VillageManagerSystem.cpp
@@ -121,7 +121,7 @@ int32 UVillageManagerSystem::FindCompetitorDataArrIdx(int32 AreaID)
 
 void UVillageManagerSystem::UpdatePlayerBistroRating(float Rating)
 {
-	float UpdatedRating = ((PlayerBistroRating * (PlayerBistroTotalCust - 1)) + Rating) / PlayerBistroTotalCust;
+	float UpdatedRating = ((PlayerBistroRating * (PlayerBistroTotalCust - 1)) + Rating) / (float)PlayerBistroTotalCust;
 	PlayerBistroRating = FMath::RoundToFloat(UpdatedRating * 10.0f) / 10.0f;
 }
